backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-dotenv>=1.0.0
motor>=3.6.0
pymongo[srv]>=4.9.0
openai>=1.50.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
google-generativeai>=0.8.0
certifi>=2024.0.0
groq>=0.11.0
